requirements.txt
streamlit
praw
pandas
requests
python-dotenv
python-docx
pdfplumber
openpyxl
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
async-timeout==5.0.1
attrs==25.3.0
blinker==1.9.0
blis==1.2.1
cachetools==6.2.0
catalogue==2.0.10
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
cloudpathlib==0.22.0
colorama==0.4.6
confection==0.1.5
contourpy==1.3.2
cycler==0.12.1
cymem==2.0.11
datasets==4.0.0
dill==0.3.8
et_xmlfile==2.0.0
filelock==3.19.1
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2025.3.0
gensim==4.3.3
gitdb==4.0.12
GitPython==3.1.45
huggingface-hub==0.34.4
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.16
murmurhash==1.0.13
narwhals==2.4.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
pillow==11.3.0
plotly==6.3.0
preshed==3.0.10
propcache==0.3.2
protobuf==6.32.0
pyarrow==21.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
rich==14.1.0
rpds-py==0.27.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.13.1
seaborn==0.13.2
shellingham==1.5.4
six==1.17.0
smart_open==7.3.1
smmap==5.0.2
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
streamlit==1.49.1
sympy==1.14.0
tenacity==9.1.2
thinc==8.3.4
threadpoolctl==3.6.0
tokenizers==0.22.0
toml==0.10.2
torch==2.8.0
tornado==6.5.2
tqdm==4.67.1
transformers==4.56.1
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wasabi==1.1.3
watchdog==6.0.0
weasel==0.4.1
wordcloud==1.9.4
wrapt==1.17.3
xxhash==3.5.0
yarl==1.20.1